refactor: category repository 구조 변경
