+++ src/main/java/com/catcher/batch/core/database/CategoryRepository.java
@@ -1,0 +1,11 @@
+package com.catcher.batch.core.database;
+
+import com.catcher.batch.core.domain.entity.Category;
+
+import java.util.Optional;
+
+public interface CategoryRepository {
+    Optional<Category> findByName(String name);
+
+    Category save(Category category);
+}
